simple backdoor attack works

+++ src/utils.py
@@ -241,7 +241,7 @@ def process_control():
     cfg['stats'] = make_stats()
     cfg['attack_mode'] = cfg['backdoor']['attack']
     cfg['num_attackers'] = int(cfg['backdoor']['num_attackers'])
-    cfg['poison_percent'] = int(cfg['backdoor']['poison_percent'])
+    cfg['poison_percent'] = float(cfg['backdoor']['poison_percent'])
     cfg['mark_path'] = cfg['mark']['mark_path']
     return
